app/requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.31
asyncpg==0.28.0
psycopg==3.1.10
psycopg-binary==3.2.4
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
sqlmodel==0.0.22
starlette==0.45.3
fastapi==0.115.8
uvicorn==0.34.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1
pika==1.3.2
jinja2==3.1.5
pydantic[email]==2.3.0
markdown==3.7

requests==2.32.3
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.25.3
httpx==0.28.1
black==25.1.0
flake8==7.1.2
mypy==1.15.0